renaming electron-api.ts to dialog.ts
