+++ electron/dialog.ts
